fix(auth): add auto-login for Vercel preview deployments
Enable automatic credential-based authentication in preview
environments so PR reviewers can skip manual login.
